Fix large images still not being openable
5120x1440 jxr HDR images fail to be opened when the app is in a maxed out window, because the histogram still fails to draw (EndDraw() has an exception within itself, and returns error D1225).
I haven't really spent time analyzing what the histogram is used for, but I have found that the jxr image is visualized exactly the same way whether the EndDraw() failed or not (thus the max luminance remains n/a).
I can manage to open these images, but only if the app started in a small window. If it's maxed out, it will throw the error.

I'm on the latest version of Win 11 and have a recent high end Nvidia card (I don't think it matters).
My resolution is set to 5120x1440 and DPI to 100% (so 96).

I have found that changing the "SetDpi()" call from 96 to 1 avoids the exception to be throw, though it actually kind of breaks the final image rendering.
I know that makes no sense, I was just messing around.

Before the code was changed to add the "SetDpi()" workaround, it errored out even more often, though I'm not sure exactly what improved.